feat(output): refactor to prepare for multi trx output
