+++ src/utils.ts
@@ -1,0 +1,21 @@
+import os from 'os';
+import uuid from 'uuid';
+
+import { IDSimpleType } from './trxModel';
+
+export const computerName: string = os.hostname();
+
+export const userName: string = os.userInfo().username;
+
+export const runUser = `${computerName}\\${userName}`;
+
+export function createUuid(): IDSimpleType {
+  return uuid.v4() as IDSimpleType;
+}
+
+// generated by https://www.uuidgenerator.net/
+const PW_TRX_REPORTER_NAMESPACE = 'e2d7181b-0942-4431-a44c-ad5686b2a530';
+
+export function convertPwId2Uuid(id: string): IDSimpleType {
+  return uuid.v5(id, PW_TRX_REPORTER_NAMESPACE) as IDSimpleType;
+}
